Move load_ssm_parameters function to aws_functions module and update import in lease_dataframe.py
